+++ lib/src/devices/src/variants.dart
@@ -1,0 +1,12 @@
+part of '../devices.dart';
+
+enum DeviceVariant {
+  /// [standard] is the standard variant of the device.
+  standard,
+
+  /// [ats] is the variant of the device for ATS.
+  ats,
+
+  /// [mappit] is the variant of the device for Mappit.
+  mappit,
+}